Separate class for docloc metadata saving
